Revert centrifuge per-vendor folder refactor
Hold off on the agilent/ subpackage split (and its deprecation shims) until
the broader machine-interface architecture lands (see #1000). This PR now
only adds the new HighRes MicroSpin files; existing VSpin imports are
unchanged.
