+++ docs/user_guide/01_material-handling/centrifuge/_centrifuge.md
@@ -12,7 +12,8 @@ The {class}`~pylabrobot.centrifuge.centrifuge.Centrifuge` class has a number of 
 - {meth}`~pylabrobot.centrifuge.centrifuge.Centrifuge.unlock_bucket`: Unlock centrifuge buckets.
 - {meth}`~pylabrobot.centrifuge.centrifuge.Centrifuge.go_to_bucket1`: Rotate to Bucket 1.
 - {meth}`~pylabrobot.centrifuge.centrifuge.Centrifuge.go_to_bucket2`: Rotate to Bucket 2.
-- {meth}`~pylabrobot.centrifuge.centrifuge.Centrifuge.spin`: Start a spin cycle. (The older {meth}`~pylabrobot.centrifuge.centrifuge.Centrifuge.start_spin_cycle` method is a deprecated alias.)
+- {meth}`~pylabrobot.centrifuge.centrifuge.Centrifuge.rotate_distance`: Rotate the buckets a specified distance (8000 = 360 degrees).
+- {meth}`~pylabrobot.centrifuge.centrifuge.Centrifuge.start_spin_cycle`: Start centrifuge spin cycle.
 
 PLR supports the following centrifuges:
