+++ pylabrobot/centrifuge/__init__.py
@@ -1,4 +1,4 @@
-from .agilent import Access2, Access2Backend, VSpinBackend
+from .access2 import Access2
 from .centrifuge import Centrifuge, Loader
 from .highres import (
   MicroSpin,
@@ -13,24 +13,4 @@ from .standard import (
   LoaderNoPlateError,
   NotAtBucketError,
 )
-
-__all__ = [
-  # Front-end
-  "Centrifuge",
-  "Loader",
-  # Standard / errors
-  "BucketHasPlateError",
-  "BucketNoPlateError",
-  "CentrifugeDoorError",
-  "LoaderNoPlateError",
-  "NotAtBucketError",
-  # Agilent (VSpin + Access2 loader)
-  "Access2",
-  "Access2Backend",
-  "VSpinBackend",
-  # HighRes Biosolutions (MicroSpin)
-  "MicroSpin",
-  "MicroSpinBackend",
-  "MicroSpinError",
-  "MicroSpinProtocolError",
-]
+from .vspin_backend import Access2Backend, VSpinBackend
