--- pylabrobot/centrifuge/agilent/__init__.py
@@ -1,8 +1,0 @@
-from .access2 import Access2
-from .vspin_backend import Access2Backend, VSpinBackend
-
-__all__ = [
-  "Access2",
-  "Access2Backend",
-  "VSpinBackend",
-]
